Fix missing license header and use Basic instead of Base
